+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   - Removed `onUpdate` configuration option.  If you want to re-generate a slug
     on update, then set the model's slug to `null` before saving.  Otherwise, existing
     slugs will never be overwritten.
+  - `createSlug()` is no longer a static method on the model, but is a public method
+    on the _SlugService_ class, with a different method signature (see docs)
   
 
 ## 3.1.4 - 03-Jan-2016